+++ app/src/main/cpp/pcsx2/pcsx2/R5900Def.h
@@ -111,8 +111,8 @@ struct cpuRegisters {
     u32 code;			// current instruction
     PERFregs PERF;
     u32 eCycle[32];
-    u32 sCycle[32];		// for internal counters
-    u32 cycle;			// calculate cpucycles..
+    u64 sCycle[32];		// for internal counters
+    u64 cycle;			// calculate cpucycles..
     u32 interrupt;
     int branch;
     int opmode;			// operating mode
@@ -121,10 +121,10 @@ struct cpuRegisters {
     u32 pcWriteback;
 
     // if cpuRegs.cycle is greater than this cycle, should check cpuEventTest for updates
-    u32 nextEventCycle;
-    u32 lastEventCycle;
-    u32 lastCOP0Cycle;
-    u32 lastPERFCycle[2];
+    u64 nextEventCycle;
+    u64 lastEventCycle;
+    u64 lastCOP0Cycle;
+    u64 lastPERFCycle[2];
 };
 
 union FPRreg {
